Remove unused social server
